+++ lib/presentation/favori_heart_view_model.dart
@@ -1,6 +1,6 @@
 import 'package:equatable/equatable.dart';
 import 'package:pass_emploi_app/redux/states/app_state.dart';
-import 'package:pass_emploi_app/redux/states/offre_emploi_favoris_state.dart';
+import 'package:pass_emploi_app/redux/states/offre_emploi_favoris_id_state.dart';
 import 'package:redux/redux.dart';
 
 class FavoriHeartViewModel extends Equatable {
