+++ lib/redux/states/app_state.dart
@@ -12,7 +12,7 @@ import 'package:pass_emploi_app/redux/states/user_action_state.dart';
 import 'package:pass_emploi_app/redux/states/user_action_update_state.dart';
 
 import 'login_state.dart';
-import 'offre_emploi_favoris_state.dart';
+import 'offre_emploi_favoris_id_state.dart';
 import 'offre_emploi_search_results_state.dart';
 import 'offre_emploi_search_state.dart';
