+++ test/feature/offre_emploi_favoris_test.dart
@@ -2,20 +2,18 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:pass_emploi_app/models/user.dart';
 import 'package:pass_emploi_app/redux/actions/ui_actions.dart';
 import 'package:pass_emploi_app/redux/states/app_state.dart';
-import 'package:pass_emploi_app/redux/states/offre_emploi_favoris_state.dart';
+import 'package:pass_emploi_app/redux/states/offre_emploi_favoris_id_state.dart';
 import 'package:pass_emploi_app/repositories/offre_emploi_favoris_repository.dart';
 import 'package:pass_emploi_app/repositories/user_repository.dart';
 
 import '../doubles/dummies.dart';
 import '../utils/test_setup.dart';
+import 'package:redux/src/store.dart';
 
 main() {
   test("favoris id should be loaded at app startup when logged in", () async {
     // Given
-    final testStoreFactory = TestStoreFactory();
-    testStoreFactory.offreEmploiFavorisRepository = OffreEmploiFavorisRepositorySuccessStub();
-    testStoreFactory.userRepository = UserRepositoryLoggedInStub();
-    final store = testStoreFactory.initializeReduxStore(initialState: AppState.initialState());
+    Store<AppState> store = _successStore();
 
     final successState =
         store.onChange.firstWhere((element) => element.offreEmploiFavorisState is OffreEmploiFavorisIdLoadedState);
@@ -28,6 +26,28 @@ main() {
     final favorisState = (loadedFavoris.offreEmploiFavorisState as OffreEmploiFavorisIdLoadedState);
     expect(favorisState.offreEmploiFavorisListId, ["1", "2", "4"]);
   });
+
+  test("favori state should be updated when favori is added and api call succeeds", () {
+    // Given
+    Store<AppState> store = _successStore();
+
+    final successState =
+    store.onChange.firstWhere((element) => element.offreEmploiFavorisState is OffreEmploiFavorisIdLoadedState);
+
+    // When
+
+    // Then
+
+
+  });
+}
+
+Store<AppState> _successStore() {
+  final testStoreFactory = TestStoreFactory();
+  testStoreFactory.offreEmploiFavorisRepository = OffreEmploiFavorisRepositorySuccessStub();
+  testStoreFactory.userRepository = UserRepositoryLoggedInStub();
+  final store = testStoreFactory.initializeReduxStore(initialState: AppState.initialState());
+  return store;
 }
 
 class OffreEmploiFavorisRepositorySuccessStub extends OffreEmploiFavorisRepository {
